🚚Move all from fela folder and get rid of dependencies.js
