+++ gatsby-theme-fela/gatsby-browser.js
@@ -1,6 +1,6 @@
 const React = require("react")
 const { wrapWithFelaRenderer } = require("./render-helpers")
-const FelaProvider = require("./src/fela/FelaProvider").default
+const FelaProvider = require("./src/components/FelaProvider").default
 
 exports.wrapRootElement = ({ element }) => {
   const { wrapped } = wrapWithFelaRenderer(
