+++ gatsby-theme-fela/gatsby-ssr.js
@@ -2,7 +2,7 @@ const React = require("react")
 const { renderToString } = require("react-dom/server")
 const { renderToSheetList } = require("fela-dom")
 const { wrapWithFelaRenderer } = require("./render-helpers")
-const FelaProvider = require("./src/fela/FelaProvider").default
+const FelaProvider = require("./src/components/FelaProvider").default
 
 exports.wrapRootElement = ({ element }) => {
   return <FelaProvider>{element}</FelaProvider>
