+++ gatsby-theme-fela/render-helpers.js
@@ -1,7 +1,7 @@
 const React = require("react")
 const { createRenderer } = require("fela")
 const { RendererProvider } = require("react-fela")
-const Config = require("./src/fela/config")
+const Config = require("./src/config")
 
 exports.wrapWithFelaRenderer = element => {
   const renderer = createRenderer(Config.rendererConfig)
