--- gatsby-theme-fela/src/fela/config/enhancers.js
@@ -1,10 +1,0 @@
-import { felaMonolithic } from '../dependencies';
-
-const enhancers = [
-    process.env.NODE_ENV === 'development' &&
-        felaMonolithic({
-            prettySelectors: true,
-        }),
-].filter(Boolean);
-
-export default enhancers;
