--- gatsby-theme-fela/src/fela/dependencies.js
@@ -1,18 +1,0 @@
-import * as Consts from '../constants';
-import { default as styles, theme } from '../styles';
-
-export { Consts, styles, theme };
-
-export { default as React, PureComponent } from 'react';
-export { default as PropTypes } from 'prop-types';
-
-export { createRenderer } from 'fela';
-export { RendererProvider, ThemeProvider } from 'react-fela';
-
-export { default as felaMonolithic } from 'fela-monolithic';
-export { default as felaPluginValidator } from 'fela-plugin-validator';
-export { default as felaPluginUnit } from 'fela-plugin-unit';
-export { default as felaPluginNamedKeys } from 'fela-plugin-named-keys';
-export { default as felaPluginFriendlyPseudoClass } from 'fela-plugin-friendly-pseudo-class';
-export { default as felaPluginExpandShorthand } from 'fela-plugin-expand-shorthand';
-
